# src/main/resources/apidocs/openapi.yaml
openapi: 3.0.1
info:
  title: Einsatzleitplatz
  version: 0.0.1
servers:
- url: http://localhost
  description: Generated server url
security:
- Basic: []
paths:
  /users/login:
    post:
      tags:
      - user-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /users:
    get:
      tags:
      - user-controller
      operationId: getUsers
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          minimum: 0
          type: string
          default: "0"
      - name: limit
        in: query
        required: false
        schema:
          maximum: 200
          minimum: 1
          type: string
          default: "50"
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
  /users/{id}:
    get:
      tags:
      - user-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    Login:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    User:
      required:
      - username
      type: object
      properties:
        username:
          type: string
...
    Pagination:
      required:
      - limit
      - offset
      - totalElements
      type: object
      properties:
        offset:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int32
        currentPage:
          type: integer
          format: int32
        lastPage:
          type: boolean
        totalPages:
          type: integer
          format: int32
    UserList:
      required:
      - data
      - pagination
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
  securitySchemes:
    Basic:
      type: http
      scheme: basic